_archive/guide-cortex-anthropic-redirect/requirements.txt
anthropic>=0.42.0
cryptography>=43.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
